+++ lib/src/utils.dart
@@ -9,11 +9,9 @@ extension RectX on Rect {
     return borderRadius.toRRect(this);
   }
 
-  /// Whether the provided rect is inside this rect
-  bool containsRect(Rect rect) {
-    if (rect.left < left || rect.right > right) return false;
-    if (rect.top < top || rect.bottom > bottom) return false;
-    return true;
+  /// Whether the given [rect] intersects with this rect
+  bool intersectsWith(Rect rect) {
+    return contains(rect.topLeft) || contains(rect.topRight) || contains(rect.bottomLeft) || contains(rect.bottomRight);
   }
 }
 
@@ -23,14 +21,28 @@ extension RenderObjectX on RenderObject {
   bool get hasParentData => parentData.runtimeType != ParentData;
 
   /// The widget that built this render object
-  Widget? get widget => debugCreator is DebugCreator
-      ? (debugCreator as DebugCreator).element.widget
-      : null;
+  Widget? get widget => debugCreator is DebugCreator ? (debugCreator as DebugCreator).element.widget : null;
 
   /// finds the first parent object the matches the given name
   AbstractNode? findParentWithName(String name) {
     AbstractNode? find(AbstractNode box) {
+       if(box is RenderSemanticsAnnotations) {
+         print(box.container);
+       }
       if (box.runtimeType.toString() == name) {
+        return box;
+      } else if (box.parent != null) {
+        return find(box.parent!);
+      }
+      return null;
+    }
+
+    return find(this);
+  }
+
+  RenderSemanticsAnnotations? findFirstAnnoation() {
+    RenderSemanticsAnnotations? find(AbstractNode box) {
+      if (box is RenderSemanticsAnnotations) {
         return box;
       } else if (box.parent != null) {
         return find(box.parent!);