+++ CHANGELOG.md
@@ -1,13 +1,27 @@
+## 0.3.0 [Breaking Change]
+- Fix crash when running in release mode.
+- ColoredBox and Container with non-null color will need now need to be exclusively wrapped with a
+  Skeleton.coloredBox annotation widget due to platform limitations.
+
 ## 0.2.0
+
 - Add SliverSkeletonizer and Skeletonizer.sliver to handle sliver widgets.
+
 ## 0.1.2
+
 - improve docs
+
 ## 0.1.1
+
 - Fix containers with no descendants should not be ignored
 - Add tutorial article link to documentation
+
 ## 0.1.0+2
+
 - fix docs formatting.
 - ## 0.1.0+1
 - add docs.
+
 ## 0.1.0
+
 - initial release.